test: split test_build.py into test_build_{disk,iso}
Split this big test into smaller files because we
will run the tests in parallel via a dynamic (per-file)
github matrix. This will allow faster tests and easier
re-runs if a single test is flaky only a small subset
will have to be retriggered.
