+++ test/test_build_disk.py
@@ -610,53 +610,6 @@ def test_image_build_without_se_linux_denials(image_type):
         f"denials in log {image_type.journal_output}"
 
 
-@pytest.mark.skipif(platform.system() != "Linux", reason="boot test only runs on linux right now")
-@pytest.mark.parametrize("image_type", gen_testcases("anaconda-iso"), indirect=["image_type"])
-def test_iso_installs(image_type):
-    installer_iso_path = image_type.img_path
-    test_disk_path = installer_iso_path.with_name("test-disk.img")
-    with open(test_disk_path, "w", encoding="utf8") as fp:
-        fp.truncate(10_1000_1000_1000)
-    # install to test disk
-    with QEMU(test_disk_path, cdrom=installer_iso_path) as vm:
-        vm.start(wait_event="qmp:RESET", snapshot=False, use_ovmf=True)
-        vm.force_stop()
-    # boot test disk and do extremly simple check
-    with QEMU(test_disk_path) as vm:
-        vm.start(use_ovmf=True)
-        exit_status, _ = vm.run("true", user=image_type.username, password=image_type.password)
-        assert exit_status == 0
-        assert_kernel_args(vm, image_type)
-
-
-def osinfo_for(it: ImageBuildResult, arch: str) -> str:
-    base = "Media is an installer for OS"
-    if it.container_ref.endswith("/centos-bootc/centos-bootc:stream9"):
-        return f"{base} 'CentOS Stream 9 ({arch})'\n"
-    if it.container_ref.endswith("/centos-bootc/centos-bootc:stream10"):
-        return f"Media is an installer for OS 'CentOS Stream 10 ({arch})'\n"
-    if "/fedora/fedora-bootc:" in it.container_ref:
-        ver = it.container_ref.rsplit(":", maxsplit=1)[1]
-        return f"{base} 'Fedora Server {ver} ({arch})'\n"
-    raise ValueError(f"unknown osinfo string for '{it.container_ref}'")
-
-
-@pytest.mark.skipif(platform.system() != "Linux", reason="osinfo detect test only runs on linux right now")
-@pytest.mark.parametrize("image_type", gen_testcases("anaconda-iso"), indirect=["image_type"])
-def test_iso_os_detection(image_type):
-    installer_iso_path = image_type.img_path
-    arch = image_type.img_arch
-    if not arch:
-        arch = platform.machine()
-    result = subprocess.run([
-        "osinfo-detect",
-        installer_iso_path,
-    ], capture_output=True, text=True, check=True)
-    osinfo_output = result.stdout
-    expected_output = f"Media is bootable.\n{osinfo_for(image_type, arch)}"
-    assert osinfo_output == expected_output
-
-
 @pytest.mark.skipif(platform.system() != "Linux", reason="osinfo detect test only runs on linux right now")
 @pytest.mark.skipif(not testutil.has_executable("unsquashfs"), reason="need unsquashfs")
 @pytest.mark.parametrize("image_type", gen_testcases("anaconda-iso"), indirect=["image_type"])
