+++ test/test_build_iso.py
@@ -1,0 +1,85 @@
+import os
+import platform
+import subprocess
+from contextlib import ExitStack
+
+import pytest
+# local test utils
+import testutil
+from containerbuild import build_container_fixture    # pylint: disable=unused-import
+from testcases import gen_testcases
+from vm import QEMU
+
+from test_build_disk import (
+    assert_kernel_args,
+    ImageBuildResult,
+)
+from test_build_disk import (  # pylint: disable=unused-import
+    gpg_conf_fixture,
+    image_type_fixture,
+    registry_conf_fixture,
+    shared_tmpdir_fixture,
+)
+
+
+@pytest.mark.skipif(platform.system() != "Linux", reason="boot test only runs on linux right now")
+@pytest.mark.parametrize("image_type", gen_testcases("anaconda-iso"), indirect=["image_type"])
+def test_iso_installs(image_type):
+    installer_iso_path = image_type.img_path
+    test_disk_path = installer_iso_path.with_name("test-disk.img")
+    with open(test_disk_path, "w", encoding="utf8") as fp:
+        fp.truncate(10_1000_1000_1000)
+    # install to test disk
+    with QEMU(test_disk_path, cdrom=installer_iso_path) as vm:
+        vm.start(wait_event="qmp:RESET", snapshot=False, use_ovmf=True)
+        vm.force_stop()
+    # boot test disk and do extremly simple check
+    with QEMU(test_disk_path) as vm:
+        vm.start(use_ovmf=True)
+        exit_status, _ = vm.run("true", user=image_type.username, password=image_type.password)
+        assert exit_status == 0
+        assert_kernel_args(vm, image_type)
+
+
+def osinfo_for(it: ImageBuildResult, arch: str) -> str:
+    base = "Media is an installer for OS"
+    if it.container_ref.endswith("/centos-bootc/centos-bootc:stream9"):
+        return f"{base} 'CentOS Stream 9 ({arch})'\n"
+    if it.container_ref.endswith("/centos-bootc/centos-bootc:stream10"):
+        return f"Media is an installer for OS 'CentOS Stream 10 ({arch})'\n"
+    if "/fedora/fedora-bootc:" in it.container_ref:
+        ver = it.container_ref.rsplit(":", maxsplit=1)[1]
+        return f"{base} 'Fedora Server {ver} ({arch})'\n"
+    raise ValueError(f"unknown osinfo string for '{it.container_ref}'")
+
+
+@pytest.mark.skipif(platform.system() != "Linux", reason="osinfo detect test only runs on linux right now")
+@pytest.mark.parametrize("image_type", gen_testcases("anaconda-iso"), indirect=["image_type"])
+def test_iso_os_detection(image_type):
+    installer_iso_path = image_type.img_path
+    arch = image_type.img_arch
+    if not arch:
+        arch = platform.machine()
+    result = subprocess.run([
+        "osinfo-detect",
+        installer_iso_path,
+    ], capture_output=True, text=True, check=True)
+    osinfo_output = result.stdout
+    expected_output = f"Media is bootable.\n{osinfo_for(image_type, arch)}"
+    assert osinfo_output == expected_output
+
+
+@pytest.mark.skipif(platform.system() != "Linux", reason="osinfo detect test only runs on linux right now")
+@pytest.mark.skipif(not testutil.has_executable("unsquashfs"), reason="need unsquashfs")
+@pytest.mark.parametrize("image_type", gen_testcases("anaconda-iso"), indirect=["image_type"])
+def test_iso_install_img_is_squashfs(tmp_path, image_type):
+    installer_iso_path = image_type.img_path
+    with ExitStack() as cm:
+        mount_point = tmp_path / "cdrom"
+        mount_point.mkdir()
+        subprocess.check_call(["mount", installer_iso_path, os.fspath(mount_point)])
+        cm.callback(subprocess.check_call, ["umount", os.fspath(mount_point)])
+        # ensure install.img is the "flat" squashfs, before PR#777 the content
+        # was an intermediate ext4 image "squashfs-root/LiveOS/rootfs.img"
+        output = subprocess.check_output(["unsquashfs", "-ls", mount_point / "images/install.img"], text=True)
+        assert "usr/bin/bootc" in output
